# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API para Importação de Arquivo CNAB",
        "description": "A API tem o objetivo de possiblitar a importação de arquivo CNAB com os dados das movimentações finanaceira de várias lojas e exibir os lançamentos",
        "version": "0.1"
    },
    "paths": {
        "/api/transaction": {
            "get": {
                "tags": [
                    "Transações"
                ],
                "summary": "Lista as Transações",
                "description": "Endpoint destinado a listagem das Transações",
                "operationId": "listTransacoes",
                "responses": {
                    "200": {
                        "description": "Lista de transações",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/summary": {
            "get": {
                "tags": [
                    "Transações"
                ],
                "summary": "Lista as Transações",
                "description": "Endpoint destinado a prover informações gerais de dados importados",
                "operationId": "summaryTransacoes",
                "responses": {
                    "200": {
                        "description": "Agrupamento de infomrações já importadas",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Resultado será um objeto",
                                        "value": {
                                            "data": {
                                                "totImports": 4,
                                                "totTransactions": 64,
                                                "totStores": 6,
                                                "totCredit": 6387.11,
                                                "totDebit": 24906,
                                                "totBalance": -18518.89
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/transaction/import": {
            "post": {
                "tags": [
                    "Transações"
                ],
                "summary": "Importação das Transações",
                "description": "Endpoint destinado a importação do arquivo CNAB com as transações das lojas",
                "operationId": "importTransacoes",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "file": {
                                        "description": "arquivo CNAB"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Importação efetuada com sucesso",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "422": {
                        "description": "Parametros esperados não foram encontrados",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        }
    }
}